requirements.txt
google-generativeai
python-dotenv
alpaca-py
